refactor: moving functions and files
