+++ src/scripts/content/uiEnhancer/modals/launchModalIntegration.js
@@ -1,0 +1,7 @@
+import {appendModalContainer} from "./actions/appendModalContainer";
+import {displayInstallModals} from "./actions/displayInstallModals";
+
+export async function launchModalIntegration(domain) {
+  await appendModalContainer();
+  await displayInstallModals(domain);
+}
